fix: validator list empty state and dividers
